+++ all_tx.sh
@@ -7,6 +7,16 @@ read -p ">> " PLATFORM
 echo "Enter the number of iterations (e.g., 10):"
 read -p ">> " REPEAT
 
+source ./tx_commons.sh
+echo "Current configurations: "
+echo -e "\tTest time (sec): ${TIME}"
+echo -e "\tMessage sizes (B): ${M_SIZES[@]}"
+echo "Continue? (y/n)"
+read -p ">> " ANS
+
+if [ ${ANS} != "y" ]; then
+    exit 0
+fi
 
 # Remove existing results
 sudo rm -rf net_result/${PLATFORM}/basic/*_tx_*
@@ -15,9 +25,10 @@ sudo rm -rf net_result/${PLATFORM}/basic/*_tx_*
 echo "************************************"
 echo "**** START <DEFAULT> EXPERIMENT ****"
 echo "************************************"
-# default
-./${PLATFORM}_tx.sh -r "${REPEAT}" 
+# Default
+./${PLATFORM}_tx.sh -r ${REPEAT}
 echo ""
+
 
 echo "********************************"
 echo "**** START <CPU> EXPERIMENT ****"
@@ -26,7 +37,7 @@ echo "********************************"
 for arg in 1 2 3 4
 do
 	echo "CPU: ${arg}"
-	./${PLATFORM}_tx.sh -r "${REPEAT}" -c "${arg}"
+	./${PLATFORM}_tx.sh -r ${REPEAT} -c ${arg}
 done
 echo ""
 
@@ -40,7 +51,7 @@ if [[ ${PLATFORM} == "vm" ]]; then ARGS="2G 4G 6G"; else ARGS="512m 1G 2G 4G 6G"
 for arg in ${ARGS}
 do
 	echo "MEMORY: ${arg}"
-	./${PLATFORM}_tx.sh -r "${REPEAT}" -m "${arg}"
+	./${PLATFORM}_tx.sh -r ${REPEAT} -m ${arg}
 done
 echo ""
 
@@ -52,7 +63,7 @@ echo "***********************************"
 for arg in 1 3 5 10
 do
 	echo "STREAM: ${arg}"
-	./${PLATFORM}_tx.sh -r "${REPEAT}" -s "${arg}"
+	./${PLATFORM}_tx.sh -r ${REPEAT} -s ${arg}
 done
 echo ""
 
@@ -66,7 +77,7 @@ if [[ ${PLATFORM} == "vm" ]]; then ARGS="1 2 3"; else ARGS="1 2 3 4 8"; fi
 for arg in ${ARGS}
 do
 	echo "CONCURRENCY: ${arg}"
-	./${PLATFORM}_tx.sh -r "${REPEAT}" -n "${arg}"
+	./${PLATFORM}_tx.sh -r ${REPEAT} -n ${arg}
 done
 echo ""
